fix: some issues
